Move components to /src
