+++ packages/editor/src/components/Section.js
